+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 2.0.25
 ++++++
+* Adding paging support for storage list commands.
 * Minor fixes.
 
 2.0.24